move colorPaletteDefaults and colorPaletteParam to utilities
